+++ demo/symfony8/config/bundles.php
@@ -1,13 +1,20 @@
 <?php
 
 declare(strict_types=1);
+use Nowo\SelectAllChoiceBundle\NowoSelectAllChoiceBundle;
+use Nowo\TwigInspectorBundle\NowoTwigInspectorBundle;
+use Symfony\Bundle\DebugBundle\DebugBundle;
+use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
+use Symfony\Bundle\TwigBundle\TwigBundle;
+use Symfony\Bundle\WebProfilerBundle\WebProfilerBundle;
+use Symfony\UX\StimulusBundle\StimulusBundle;
 
 return [
-    Symfony\Bundle\FrameworkBundle\FrameworkBundle::class       => ['all' => true],
-    Symfony\Bundle\TwigBundle\TwigBundle::class                 => ['all' => true],
-    Symfony\Bundle\DebugBundle\DebugBundle::class               => ['dev' => true, 'test' => true],
-    Symfony\Bundle\WebProfilerBundle\WebProfilerBundle::class   => ['dev' => true, 'test' => true],
-    Symfony\UX\StimulusBundle\StimulusBundle::class             => ['all' => true],
-    Nowo\SelectAllChoiceBundle\NowoSelectAllChoiceBundle::class => ['all' => true],
-    Nowo\TwigInspectorBundle\NowoTwigInspectorBundle::class     => ['dev' => true, 'test' => true],
+    FrameworkBundle::class           => ['all' => true],
+    TwigBundle::class                => ['all' => true],
+    DebugBundle::class               => ['dev' => true, 'test' => true],
+    WebProfilerBundle::class         => ['dev' => true, 'test' => true],
+    StimulusBundle::class            => ['all' => true],
+    NowoSelectAllChoiceBundle::class => ['all' => true],
+    NowoTwigInspectorBundle::class   => ['dev' => true, 'test' => true],
 ];